+++ src/TagLib/IFD/Entries/UndefinedIFDEntry.cs
@@ -29,10 +29,10 @@ namespace TagLib.IFD.Entries
 	/// </summary>
 	public class UndefinedIFDEntry : IFDEntry
 	{
-		public uint Tag { get; private set; }
+		public ushort Tag { get; private set; }
 		public ByteVector Data { get; private set; }
 
-		public UndefinedIFDEntry (uint tag, ByteVector data)
+		public UndefinedIFDEntry (ushort tag, ByteVector data)
 		{
 			Tag = tag;
 			Data = data;
